feat(compiler): export custom types in compiled output (#3710)
NOTE: This was originally introduced as a feature in Stencil 2.x, but was reverted when a breaking change was identified.

Initial PR (orginally merged as a part of Stencil 2.x): #3612
Reversion of above PR (due to identified breaking change): #3708

Types for custom events are now re-exported from the `components.d.ts` file. This allows those types to be easily accessed from the root of the type distribution (so long as the `index.ts` file is set to export all definitions from the `components.d.ts` file):
`import { MyCustomType } from '@my-lib/types';`

Or, when using `dist-custom-elements`, these types can now be accessed from the custom element output:
`import { MyCustomType } from '@my-custom-elements-output';`

BREAKING CHANGE: Any project that currently manually maintains a means of creating type definitions for types created as a part of the Stencil project will either simply need to remove that manual step from their development or update it to re-export the types exported from the generated `component.d.ts` file