refactor(build): use 'paths' set in TS config in alias plugin
This refactors our 'alias' rollup plugin to determine what paths to
resolve things to based on the settings we have in `tsconfig.json`.

This is relevant for the mappings that we have set in `paths`, where we
map things like `@utils` to `src/utils/index.ts`. TypeScript, because
reasons, does not provide an option to automatically resolve those paths
to normal relative imports when compiling.

Our solution to this at present is to maintain a list of modules which
should be rewritten in this `alias` plugin. However, the information
about which module names map to which files is already set in our
`tsconfig.json`! So it would be nicer if we didn't have to maintain that
information in two places.

This change modifies the `alias` plugin so that it looks up module names
in the `paths` object in `tsconfig` and then uses TypeScript's module
resolution helper function to ensure it has the right path for the
module, and then re-writes the path to point to the right compiled file
in the built output.

Note that (unfortunately) this change cannot do the same thing for our
`.d.ts` files :( We'll need to figure out another solution to hopefully
copy those automatically and re-write import paths within them to point
to the right places.